requirements_dev.txt
bandit
flake8
pylint
pylint-django~=2.3.0
django-stubs
